requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
firebase==4.0.1
firebase-admin==6.4.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
gcloud==0.18.3
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==2.7
importlib-metadata==6.11.0
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jws==0.1.3
kubernetes==29.0.0
llama-index==0.10.15
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.7
llama-index-core==0.10.15
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.6
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.5
llama-parse==0.3.5
llamaindex-py-client==0.1.13
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.0
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
posthog==3.4.2
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
PyDrive==1.3.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.0.2
PyPika==0.48.9
pyproject_hooks==1.0.0
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.23
uvicorn==0.27.1
uvloop==0.19.0
validators==0.22.0
watchdog==4.0.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
